feat : 레시피데이터컴포넌트분리
