+++ components/main/bestRecipe.tsx
@@ -2,10 +2,10 @@ import type { NextPage } from "next";
 import { useEffect, useState } from "react";
 import { query, collection, orderBy, limit, getDocs } from "firebase/firestore";
 import { dbService } from "@/config/firebase";
-import Link from "next/link";
+import RecipeData from "../search/RecipeData";
 
 const BestRecipe: NextPage = () => {
-    const [currentItems, setCurrentItems] = useState<RecipeProps[]>([]);
+    const [dataResults, setDataResults] = useState<RecipeProps[]>([]);
 
     const getList = async () => {
         const items = query(
@@ -18,7 +18,7 @@ const BestRecipe: NextPage = () => {
             ...doc.data(),
             id: doc.id,
         }));
-        setCurrentItems(newData);
+        setDataResults(newData);
     };
 
     useEffect(() => {
@@ -33,36 +33,7 @@ const BestRecipe: NextPage = () => {
                     타쿠의 식탁에서 HOT한 요즘유행요리
                 </p>
             </div>
-            <div className="grid grid-cols-3 gap-6">
-                {currentItems?.map((item) => {
-                    return (
-                        <div key={item.id} className="w-[316px]">
-                            <div className="w-full h-[188px] overflow-hidden mx-auto">
-                                <picture>
-                                    <img
-                                        src={`${item.thumbnail}`}
-                                        className="w-full h-full object-cover"
-                                        alt="recipe picture"
-                                        width={800}
-                                        height={500}
-                                    />
-                                </picture>
-                            </div>
-                            <ul className="text-sm text-slate-500 space-x-2 flex">
-                                <li>&#35;{item.animationTitle}</li>
-                                <li>&#35;{item.cookingTime}</li>
-                            </ul>
-                            <div className="text-sm text-red-400">
-                                {item.displayStatus === "회원 공개" &&
-                                    `#${item.displayStatus}`}
-                            </div>
-                            <p className="text-lg text-slate-900 font-semibold">
-                                {item.foodTitle}
-                            </p>
-                        </div>
-                    );
-                })}
-            </div>
+            <RecipeData dataResults={dataResults} />
         </>
     );
 };
