+++ pages/searchPage/index.tsx
@@ -1,4 +1,3 @@
-import SearchRecipeBar from "@/components/search/SearchRecipeBar";
 import SideCategory from "@/components/search/SideCategory";
 import { dbService } from "@/config/firebase";
 import { cls } from "@/util";
@@ -7,16 +6,13 @@ import Fuse from "fuse.js";
 import type { NextPage } from "next";
 import { useRouter } from "next/router";
 import { ChangeEvent, FormEvent, useEffect, useState } from "react";
+import RecipeData from "@/components/search/RecipeData";
 
 const SearchData: NextPage = () => {
     const router = useRouter();
     const deliverKeyword = router.query.keyword;
 
-    // 상세페이지이동
-    const goToDetail = (id: any) => {
-        router.push(`/detailRecipePage/${id}`);
-    };
-
+    // 인기순,최신순
     const [isBest, setIsBest] = useState(false);
     const activeBestBtn = () => {
         setIsBest(true);
@@ -106,7 +102,7 @@ const SearchData: NextPage = () => {
                     )}
                     onClick={activeBestBtn}
                 >
-                    추천순
+                    인기순
                 </li>
                 <li
                     className={cls(
@@ -126,45 +122,9 @@ const SearchData: NextPage = () => {
                 </p>
             ) : null}
             <div className="w-3/4 border-b border-border mb-[30px]"></div>
-            <div className="w-3/4 flex justify-center mb-20">
+            <div className="w-3/4 flex justify-start gap-7 mb-20">
                 <SideCategory />
-                <div className="grid grid-cols-3 gap-5 gap-y-14">
-                    {dataResults?.map((item) => {
-                        return (
-                            <div
-                                key={item.id}
-                                className="w-[316px] cursor-pointer"
-                                onClick={() => goToDetail(item.id)}
-                            >
-                                <div className="w-full h-[188px] overflow-hidden mx-auto relative">
-                                    {item.displayStatus === "회원 공개" && (
-                                        <div className="w-full h-full bg-slate-50 opacity-60 absolute top-0 left-0"></div>
-                                    )}
-                                    <picture>
-                                        <img
-                                            src={`${item.thumbnail}`}
-                                            className="w-full h-full object-cover"
-                                            alt="recipe picture"
-                                            width={800}
-                                            height={500}
-                                        />
-                                    </picture>
-                                </div>
-                                <ul className="text-sm text-slate-500 space-x-2 flex">
-                                    <li>&#35;{item.animationTitle}</li>
-                                    <li>&#35;{item.cookingTime}</li>
-                                </ul>
-                                <div className="text-sm text-red-400">
-                                    {item.displayStatus === "회원 공개" &&
-                                        `#${item.displayStatus}`}
-                                </div>
-                                <p className="text-lg text-slate-900 font-semibold">
-                                    {item.foodTitle}
-                                </p>
-                            </div>
-                        );
-                    })}
-                </div>
+                <RecipeData dataResults={dataResults} />
             </div>
         </div>
     );
